requirements.txt
google-genai==0.6.0
pydantic>=2.0.0
